Move shared functions from aggregate_exec.c and hybrid_execution.c into src/aggregate/aggregate_exec_common.h
